grate: move 3d ctx stuff out of grate.h into its own header
For easier navigation through the code.
