+++ src/libgrate/grate.h
@@ -30,6 +30,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+#include "grate-3d-ctx.h"
 #include "host1x.h"
 
 #define ARRAY_SIZE(a) (sizeof(a) / sizeof((a)[0]))
@@ -42,14 +43,6 @@ struct grate;
 
 #define GRATE_SINGLE_BUFFERED (0 << 0)
 #define GRATE_DOUBLE_BUFFERED (1 << 0)
-
-enum grate_3d_ctx_cull_face
-{
-	GRATE_3D_CTX_CULL_FACE_NONE  = 0,
-	GRATE_3D_CTX_CULL_FACE_FRONT = 1,
-	GRATE_3D_CTX_CULL_FACE_BACK  = 2,
-	GRATE_3D_CTX_CULL_FACE_BOTH  = 3
-};
 
 struct grate_framebuffer *grate_framebuffer_create(struct grate *grate,
 						   unsigned int width,
@@ -142,62 +135,6 @@ float grate_profile_time_elapsed(struct grate_profile *profile);
 
 struct grate_3d_ctx;
 struct grate_texture;
-
-struct grate_3d_ctx * grate_3d_alloc_ctx(struct grate *grate);
-int grate_3d_ctx_vertex_attrib_pointer(struct grate_3d_ctx *ctx,
-				       unsigned location, unsigned size,
-				       unsigned type, unsigned stride,
-				       struct host1x_bo *data_bo);
-int grate_3d_ctx_enable_vertex_attrib_array(struct grate_3d_ctx *ctx,
-					    unsigned target);
-int grate_3d_ctx_disable_vertex_attrib_array(struct grate_3d_ctx *ctx,
-					     unsigned target);
-int grate_3d_ctx_bind_render_target(struct grate_3d_ctx *ctx,
-				    unsigned target,
-				    struct host1x_pixelbuffer *pixbuf);
-int grate_3d_ctx_set_render_target_dither(struct grate_3d_ctx *ctx,
-					  unsigned target,
-					  bool enable);
-int grate_3d_ctx_enable_render_target(struct grate_3d_ctx *ctx,
-				      unsigned target);
-int grate_3d_ctx_disable_render_target(struct grate_3d_ctx *ctx,
-				       unsigned target);
-int grate_3d_ctx_bind_program(struct grate_3d_ctx *ctx,
-			      struct grate_program *program);
-int grate_3d_ctx_set_vertex_uniform(struct grate_3d_ctx *ctx,
-				    unsigned location, unsigned nb,
-				    float *values);
-int grate_3d_ctx_set_fragment_uniform(struct grate_3d_ctx *ctx,
-				      unsigned location, unsigned nb,
-				      float *value);
-void grate_3d_ctx_set_depth_range(struct grate_3d_ctx *ctx,
-				  float near, float far);
-void grate_3d_ctx_set_dither(struct grate_3d_ctx *ctx, uint32_t unk);
-void grate_3d_ctx_set_viewport_bias(struct grate_3d_ctx *ctx,
-				    float x, float y, float z);
-void grate_3d_ctx_set_viewport_scale(struct grate_3d_ctx *ctx,
-				     float width, float height, float depth);
-void grate_3d_ctx_set_point_params(struct grate_3d_ctx *ctx, uint32_t params);
-void grate_3d_ctx_set_point_size(struct grate_3d_ctx *ctx, float size);
-void grate_3d_ctx_set_line_params(struct grate_3d_ctx *ctx, uint32_t params);
-void grate_3d_ctx_set_line_width(struct grate_3d_ctx *ctx, float width);
-void grate_3d_ctx_use_guardband(struct grate_3d_ctx *ctx, bool enabled);
-void grate_3d_ctx_set_front_direction_is_cw(struct grate_3d_ctx *ctx,
-					    bool front_cw);
-void grate_3d_ctx_set_cull_face(struct grate_3d_ctx *ctx,
-                                enum grate_3d_ctx_cull_face cull_face);
-void grate_3d_ctx_set_scissor(struct grate_3d_ctx *ctx,
-			      unsigned x, unsigned width,
-			      unsigned y, unsigned height);
-void grate_3d_ctx_set_point_coord_range(struct grate_3d_ctx *ctx,
-					float min_s, float max_s,
-					float min_t, float max_t);
-void grate_3d_ctx_set_polygon_offset(struct grate_3d_ctx *ctx,
-				     float units, float factor);
-void grate_3d_ctx_set_provoking_vtx_last(struct grate_3d_ctx *ctx, bool last);
-int grate_3d_ctx_bind_texture(struct grate_3d_ctx *ctx,
-			      unsigned location,
-			      struct grate_texture *tex);
 
 void grate_3d_draw_elements(struct grate_3d_ctx *ctx,
 			    unsigned primitive_type,
